Powder Snow has 100% chance to freeze
I am testing the frostbite effect so I edited the effect of powder snow to freeze 100% of the time. This is probably going to be reduced to 30% later on but I need a test effect. Additionally, I fixed the spacing on some moves I made fairy type while I was in there